% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anki.R
\name{anki}
\alias{anki}
\title{Hide/Show text}
\usage{
anki(text)
}
\arguments{
\item{text}{character. Text to toggle.}
}
\value{
HTML code span.
}
\description{
This function inserts HTML snippet to help students memorize important concepts.
To make this function to work, you need to run \code{anki_setup()} at the end of
the markdown document.
}
