% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anki.R
\name{anki_setup}
\alias{anki_setup}
\title{Hide/Show text}
\usage{
anki_setup(
  background = "yellow",
  color = "blue",
  border = "slateblue",
  padding = 3,
  margin = 3,
  cursor = "pointer"
)
}
\arguments{
\item{background}{character. Background color of the blank box.}

\item{color}{character. Font color of the answer of the blank box.}

\item{border}{character. Border color of the blank box.}

\item{padding}{integer. Space between the answer text and the border.}

\item{margin}{integer. Left and right margin around the border.}

\item{cursor}{character.}
}
\value{
character. CSS/JavaScript code for anki texts.
}
\description{
This function inserts JavaScript code for anki texts.
This command should be run at the end of the markdown document.
}
