% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counter factory.}
\usage{
counter(label, sep = " ")
}
\arguments{
\item{label}{character. Default prefix for the counter.}

\item{sep}{character. Separator between label and counter.}
}
\value{
List of functions to set and get counters with label.
}
\description{
Counter factory.
}
\examples{
cnt <- counter("Quiz")
cnt$set("aaa")
cnt$set()
cnt$set("bbb", quiet = TRUE)
cnt$get("aaa")
cnt$get("bbb")
cnt$get("aaa", bare = TRUE)
cnt$dump()
}
