% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{includeAudio}
\alias{includeAudio}
\title{Include Audio media.}
\usage{
includeAudio(url)
}
\arguments{
\item{url}{character. URL of the audio. The media must not be gated.}
}
\value{
HTML5 audio tag is inserted into the document.
}
\description{
Include Audio media.
}
