% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeFlickr.R
\name{includeFlickr}
\alias{includeFlickr}
\title{Include Flickr album}
\usage{
includeFlickr(albumUrl, title = NULL, width = 480, height = 360)
}
\arguments{
\item{albumUrl}{character. Share URL of a public album}

\item{title}{character. Default NULL. If given, overwrite the title of the album.}

\item{width}{integer. Width of the widget in pixel.}

\item{height}{integer. Height of the widget in pixel.}
}
\value{
Flickr embed code.
}
\description{
Include Flickr album
}
