% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeMedia.R
\name{includeGraphics}
\alias{includeGraphics}
\title{Include graphic media.}
\usage{
includeGraphics(src, ...)
}
\arguments{
\item{src}{character. Path to the image.}

\item{...}{Parameters passed to \code{\link[knitr:include_graphics]{knitr::include_graphics()}}, other than path.}
}
\value{
The result of \code{\link[knitr:include_graphics]{knitr::include_graphics()}}
}
\description{
This function is a wrapper for \code{\link[knitr:include_graphics]{knitr::include_graphics()}}.
Since omuecon creates intermediate files in a temporary directory,
calling \code{\link[knitr:include_graphics]{knitr::include_graphics()}} does not include local images.
This function copies the media into the temporary directory and
calls \code{\link[knitr:include_graphics]{knitr::include_graphics()}}.
}
