% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{includeQuestion}
\alias{includeQuestion}
\title{Include Summary/Details block.}
\usage{
includeQuestion(
  summary,
  detail,
  summary_label = "Quiz",
  detail_label = "Answer",
  ...
)
}
\arguments{
\item{summary}{character. Text to show, typically, question string.}

\item{detail}{character. Texts to hide, typically, answer string.}

\item{summary_label}{character. Label for the open block.}

\item{detail_label}{character. Label for the hidden block.}

\item{...}{character. Concatenated with the detail argument.}
}
\value{
HTML5 summary/details block is inserted.
}
\description{
Include Summary/Details block.
}
