% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{includeYT}
\alias{includeYT}
\title{Include YouTube video.}
\usage{
includeYT(url, .class = "includeYT")
}
\arguments{
\item{url}{character. URL of the YouTube video.}

\item{.class}{character. The resulting iframe tag is surrounded by div block.
You can specify its class attribute to modify its styling.}
}
\value{
An embed code of a YouTube video is inserted.
}
\description{
Include YouTube video.
}
