% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_html}
\alias{moodle_html}
\title{Create HTML block with inline CSS.}
\usage{
moodle_html(file = NULL, dir = dirname(file), clip = TRUE, ...)
}
\arguments{
\item{file}{Path to the source HTML/Markdown file. If unspecified,
\code{\link[=file.choose]{file.choose()}} helps you to choose one.}

\item{dir}{Path to the output directory. If NULL, saved to the same directory
as the source data.}

\item{clip}{logical. If TRUE (default), the result will be copied to
the clipboard. When full_html is TRUE, the copying will be skipped.}

\item{...}{Paramters passed to \code{\link[=moodle_html_from_html]{moodle_html_from_html()}} and \code{\link[=moodle_html_from_md]{moodle_html_from_md()}}.}
}
\value{
character vector representing the resulting document (HTML fragment).
}
\description{
Create HTML block with inline CSS.
}
