% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_html_from_html}
\alias{moodle_html_from_html}
\title{Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.}
\usage{
moodle_html_from_html(
  file,
  dir = dirname(file),
  tag = "article",
  id = NULL,
  full_html = FALSE
)
}
\arguments{
\item{file}{character. Path to a standalone HTML file.}

\item{dir}{character. Path to the directory to save the resulting HTML file.}

\item{tag}{character. Outer-most tag for the resulting HTML snippet.}

\item{id}{character. id attribute for the outer-most tag.}

\item{full_html}{logical. If TRUE, produce complete html output,
convenient when drafting a document. When this option is enabled, tag and
id options are silently ignored.}
}
\value{
character. HTML block.
}
\description{
Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.
}
