% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_page}
\alias{moodle_page}
\alias{moodle_label}
\title{Wrappers for \code{\link[=moodle_html]{moodle_html()}}}
\usage{
moodle_page(file = NULL, dir = dirname(file), ...)

moodle_label(file = NULL, dir = dirname(file), ...)
}
\arguments{
\item{file}{character. Path to the markdown file.}

\item{dir}{character. Path to the output dir.}

\item{...}{Arguments passed on to \code{\link[=moodle_html]{moodle_html()}}}
}
\value{
character vector representing the resulting document (HTML fragment).
}
\description{
Convert markdown file to HTML fragment ready to be copy-pasted
for Moodle page or label, etc.
}
\details{
The output of \code{\link[=moodle_page]{moodle_page()}} is meant for Moodle pages. It
creates \code{article} tag block that makes the main part of a single
page. \code{\link[=moodle_label]{moodle_label()}} creates a more smaller part of the entire
Moodle sites, contained in a \code{div} block. It is used typically
for labels, in which more tighter style may be appropriate.
}
\section{Functions}{
\itemize{
\item \code{moodle_label()}: Wrapper for making snippets for Moodle labels.

}}
