% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omuecon-package.R
\docType{package}
\name{omuecon-package}
\alias{omuecon}
\alias{omuecon-package}
\title{omuecon: Helper Functions for Moodle}
\description{
A collection of functions to work with Moodle LMS.
}
\author{
\strong{Maintainer}: Kenji Sato \email{mail@kenjisato.jp}

}
\keyword{internal}
