% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_html}
\alias{moodle_html}
\title{Create HTML block with inline CSS.}
\usage{
moodle_html(file, dir = NULL, clip = TRUE, ...)
}
\arguments{
\item{file}{Path to the source HTML/Markdown file. s}

\item{dir}{Path to the output directory. If NULL, saved to the same directory
as the source data.}

\item{clip}{bool. If TRUE (default), the result will be copied to
the clipboard.}

\item{...}{Paramters passed to \code{\link[=moodle_html_from_md]{moodle_html_from_md()}}.}
}
\value{
character vector representing the resulting document (HTML fragment).
}
\description{
Create HTML block with inline CSS.
}
