% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_html_from_html}
\alias{moodle_html_from_html}
\title{Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.}
\usage{
moodle_html_from_html(file, dir = NULL, tag = "article", id = NULL)
}
\arguments{
\item{file}{character. Path to a standalone HTML file.}

\item{dir}{character. Path to the directory to save the resulting HTML file.}

\item{tag}{character. Outer-most tag for the resulting HTML snippet.}

\item{id}{character. id attribute for the outer-most tag.}
}
\value{
character. HTML block.
}
\description{
Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.
}
