% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_html.R
\name{moodle_html_from_html}
\alias{moodle_html_from_html}
\title{Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.}
\usage{
moodle_html_from_html(file, dir = NULL)
}
\arguments{
\item{file}{Path to a standalone HTML file.}

\item{dir}{Path to the directory to save the resulting HTML file.}
}
\value{
Invisibly returns a character vector.
}
\description{
Convert a standalone HTML file into an HTML block to be copied and
pasted for Moodle.
}
